Added stub implementation of SemanticFieldMapper
